TEST1: test 코드 추가 및 디버깅

+++ domain/src/main/kotlin/com/seungkyu/test/application/author/ports/input/service/AuthorServiceImpl.kt
@@ -1,24 +1,34 @@
 package com.seungkyu.test.application.author.ports.input.service
 
 import com.seungkyu.test.application.author.dto.*
+import com.seungkyu.test.application.author.exception.AuthorErrorCode
+import com.seungkyu.test.application.author.exception.AuthorException
 import com.seungkyu.test.application.author.ports.output.repository.AuthorRepository
+import com.seungkyu.test.application.book.ports.output.repository.BookRepository
 import com.seungkyu.test.domain.author.entity.Author
 import com.seungkyu.test.domain.author.service.AuthorDomainService
-import jakarta.validation.Valid
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
 class AuthorServiceImpl(
     private val authorRepository: AuthorRepository,
+    private val bookRepository: BookRepository,
     private val authorDomainService: AuthorDomainService
 ): AuthorService {
 
     @Transactional
-    override fun createAuthor(@Valid createAuthorCommand: CreateAuthorCommand): CreateAuthorResponse {
+    override fun createAuthor(createAuthorCommand: CreateAuthorCommand): CreateAuthorResponse {
+
+        val existedEmail = authorRepository.findByEmail(createAuthorCommand.email!!)
+
+        if (existedEmail.isPresent) {
+            throw AuthorException(AuthorErrorCode.AUTHOR_DUPLICATE_EMAIL)
+        }
+
         val author = authorDomainService.createAuthor(
             name = createAuthorCommand.name!!,
-            email = createAuthorCommand.email!!,
+            email = createAuthorCommand.email,
         )
 
         val savedAuthor = authorRepository.save(author)
@@ -43,13 +53,19 @@ class AuthorServiceImpl(
     @Transactional
     override fun updateAuthor(
         id: Int,
-        @Valid updateAuthorCommand: UpdateAuthorCommand): AuthorInfoResponse {
+        updateAuthorCommand: UpdateAuthorCommand): AuthorInfoResponse {
         val author = authorRepository.findById(id)
+
+        val existedEmail = authorRepository.findByEmail(updateAuthorCommand.email!!)
+
+        if(existedEmail.isPresent && existedEmail.get().id != id) {
+            throw AuthorException(AuthorErrorCode.AUTHOR_DUPLICATE_EMAIL)
+        }
 
         authorDomainService.updateAuthor(
             author = author,
             name = updateAuthorCommand.name!!,
-            email = updateAuthorCommand.email!!,
+            email = updateAuthorCommand.email,
         )
 
         val savedUser = authorRepository.save(author)
@@ -60,6 +76,8 @@ class AuthorServiceImpl(
     @Transactional
     override fun deleteAuthor(authorId: Int) {
         authorRepository.findById(authorId)
+        if(bookRepository.existsByAuthorId(authorId))
+            throw AuthorException(AuthorErrorCode.EXIST_RELATED_BOOK)
         authorRepository.deleteById(authorId)
     }
